requirements.txt
python-telegram-bot[job-queue]==20.7
easyocr==1.7.2
openpyxl==3.1.5
pandas==2.3.1
Pillow==11.3.0
python-dotenv==1.0.0
torch==2.7.1
opencv-python-headless==4.12.0.88
scipy==1.13.3
scikit-image==0.24.0 